// server/db/catalog.generated.ts
export const UNIT_ROWS = [
  {
    "code": "HS1-01",
    "label": "다항식",
    "school": "고등학교",
    "grade": "1학년"
  },
  {
    "code": "HS1-02",
    "label": "방정식과 부등식",
    "school": "고등학교",
    "grade": "1학년"
  },
  {
    "code": "HS1-03",
    "label": "경우의 수",
    "school": "고등학교",
    "grade": "1학년"
  },
  {
    "code": "HS1-04",
    "label": "도형의 방정식",
    "school": "고등학교",
    "grade": "1학년"
  },
  {
    "code": "HS1-05",
    "label": "집합과 명제",
    "school": "고등학교",
    "grade": "1학년"
  },
  {
    "code": "HS1-06",
    "label": "함수",
    "school": "고등학교",
    "grade": "1학년"
  },
  {
    "code": "MS1-01",
    "label": "수와 연산",
    "school": "중학교",
    "grade": "1학년"
  },
  {
    "code": "MS1-02",
    "label": "문자와 식·방정식",
    "school": "중학교",
    "grade": "1학년"
  },
  {
    "code": "MS1-03",
    "label": "좌표와 그래프",
    "school": "중학교",
    "grade": "1학년"
  },
  {
    "code": "MS1-04",
    "label": "기본 도형",
    "school": "중학교",
    "grade": "1학년"
  },
  {
    "code": "MS1-05",
    "label": "원과 입체",
    "school": "중학교",
    "grade": "1학년"
  },
  {
    "code": "MS1-06",
    "label": "자료의 정리",
    "school": "중학교",
    "grade": "1학년"
  },
  {
    "code": "MS2-01",
    "label": "유리수와 순환소수",
    "school": "중학교",
    "grade": "2학년"
  },
  {
    "code": "MS2-02",
    "label": "식의 계산",
    "school": "중학교",
    "grade": "2학년"
  },
  {
    "code": "MS2-03",
    "label": "부등식과 연립방정식",
    "school": "중학교",
    "grade": "2학년"
  },
  {
    "code": "MS2-04",
    "label": "일차함수",
    "school": "중학교",
    "grade": "2학년"
  },
  {
    "code": "MS2-05",
    "label": "삼각형과 사각형",
    "school": "중학교",
    "grade": "2학년"
  },
  {
    "code": "MS2-06",
    "label": "닮음",
    "school": "중학교",
    "grade": "2학년"
  },
  {
    "code": "MS2-07",
    "label": "피타고라스 정리",
    "school": "중학교",
    "grade": "2학년"
  },
  {
    "code": "MS2-08",
    "label": "확률",
    "school": "중학교",
    "grade": "2학년"
  },
  {
    "code": "MS3-01",
    "label": "제곱근과 실수",
    "school": "중학교",
    "grade": "3학년"
  },
  {
    "code": "MS3-02",
    "label": "다항식의 곱셈과 인수분해",
    "school": "중학교",
    "grade": "3학년"
  },
  {
    "code": "MS3-03",
    "label": "이차방정식",
    "school": "중학교",
    "grade": "3학년"
  },
  {
    "code": "MS3-04",
    "label": "이차함수",
    "school": "중학교",
    "grade": "3학년"
  },
  {
    "code": "MS3-05",
    "label": "피타고라스 정리",
    "school": "중학교",
    "grade": "3학년"
  },
  {
    "code": "MS3-06",
    "label": "삼각비",
    "school": "중학교",
    "grade": "3학년"
  },
  {
    "code": "MS3-07",
    "label": "원의 성질",
    "school": "중학교",
    "grade": "3학년"
  },
  {
    "code": "MS3-08",
    "label": "대푯값",
    "school": "중학교",
    "grade": "3학년"
  },
  {
    "code": "MS3-09",
    "label": "산포도",
    "school": "중학교",
    "grade": "3학년"
  },
  {
    "code": "MS3-10",
    "label": "산점도와 상관관계",
    "school": "중학교",
    "grade": "3학년"
  }
] as const;